requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coverage==5.5
crispy-bootstrap5==0.3.1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-heroku==0.3.1
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.7.0
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.19.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.14.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
